+++ FirebaseABTesting/Sources/Public/FIRExperimentController.h
@@ -14,6 +14,8 @@
 
 #import <Foundation/Foundation.h>
 
+@class ABTExperimentPayload;
+
 // Forward declaration to avoid importing into the module header
 typedef NS_ENUM(int32_t, ABTExperimentPayload_ExperimentOverflowPolicy);
 
@@ -37,8 +39,7 @@ NS_SWIFT_NAME(ExperimentController)
 /// existing in payloads are not affected, whose state and payload is preserved. This method
 /// compares whether the experiments have changed or not by their variant ID. This runs in a
 /// background queue and calls the completion handler when finished executing.
-/// @param origin         The originating service affected by the experiment, it is defined at
-///                       Firebase Analytics FIREventOrigins.h.
+/// @param origin         The originating service affected by the experiment.
 /// @param events         A list of event names to be used for logging experiment lifecycle events,
 ///                       if they are not defined in the payload.
 /// @param policy         The policy to handle new experiments when slots are full.
@@ -47,7 +48,7 @@ NS_SWIFT_NAME(ExperimentController)
 ///                       January 1970.).
 /// @param payloads       List of experiment metadata.
 /// @param completionHandler Code to be executed after experiments are updated in the background
-/// thread.
+///                       thread.
 - (void)updateExperimentsWithServiceOrigin:(NSString *)origin
                                     events:(FIRLifecycleEvents *)events
                                     policy:(ABTExperimentPayload_ExperimentOverflowPolicy)policy
@@ -60,8 +61,7 @@ NS_SWIFT_NAME(ExperimentController)
 /// existing in payloads are not affected, whose state and payload is preserved. This method
 /// compares whether the experiments have changed or not by their variant ID. This runs in a
 /// background queue..
-/// @param origin         The originating service affected by the experiment, it is defined at
-///                       Firebase Analytics FIREventOrigins.h.
+/// @param origin         The originating service affected by the experiment.
 /// @param events         A list of event names to be used for logging experiment lifecycle events,
 ///                       if they are not defined in the payload.
 /// @param policy         The policy to handle new experiments when slots are full.
@@ -69,7 +69,6 @@ NS_SWIFT_NAME(ExperimentController)
 ///                       (Timestamps are specified by the number of seconds from 00:00:00 UTC on 1
 ///                       January 1970.).
 /// @param payloads       List of experiment metadata.
-/// thread.
 - (void)updateExperimentsWithServiceOrigin:(NSString *)origin
                                     events:(FIRLifecycleEvents *)events
                                     policy:(ABTExperimentPayload_ExperimentOverflowPolicy)policy
@@ -86,6 +85,19 @@ NS_SWIFT_NAME(ExperimentController)
 /// @param payloads   List of experiment metadata.
 - (NSTimeInterval)latestExperimentStartTimestampBetweenTimestamp:(NSTimeInterval)timestamp
                                                      andPayloads:(NSArray<NSData *> *)payloads;
+
+/// Expires experiments that aren't in the list of running experiment payloads.
+/// @param origin The originating service affected by the experiment.
+/// @param payloads The list of valid, running experiments.
+- (void)validateRunningExperimentsForServiceOrigin:(NSString *)origin
+                         runningExperimentPayloads:(NSArray<ABTExperimentPayload *> *)payloads;
+
+/// Directly sets a given experiment to be active.
+/// @param experimentPayload The payload for the experiment that should be activated.
+/// @param origin The originating service affected by the experiment.
+- (void)activateExperiment:(ABTExperimentPayload *)experimentPayload
+          forServiceOrigin:(NSString *)origin;
+
 @end
 
 NS_ASSUME_NONNULL_END