+++ FirebaseInAppMessaging/Sources/Flows/FIRIAMDisplayExecutor.m
@@ -26,6 +26,8 @@
 #import "FIRInAppMessaging.h"
 #import "FIRInAppMessagingRenderingPrivate.h"
 
+#import <FirebaseABTesting/FIRExperimentController.h>
+
 @implementation FIRIAMDisplaySetting
 @end
 
@@ -207,6 +209,13 @@
                   @"impressionDetected called but "
                    "there is no current message ID.");
     return;
+  }
+
+  // If this is an experimental FIAM, activate the experiment.
+  if (inAppMessage.campaignInfo.experimentPayload) {
+    [[FIRExperimentController sharedInstance]
+        activateExperiment:inAppMessage.campaignInfo.experimentPayload
+          forServiceOrigin:@"fiam"];
   }
 
   if (!_currentMsgBeingDisplayed.isTestMessage) {
@@ -438,6 +447,7 @@
   FIRInAppMessagingCardDisplay *cardMessage = [[FIRInAppMessagingCardDisplay alloc]
         initWithMessageID:renderData.messageID
              campaignName:renderData.name
+        experimentPayload:definition.experimentPayload
       renderAsTestMessage:definition.isTestMessage
               triggerType:triggerType
                 titleText:title
@@ -468,6 +478,7 @@
   FIRInAppMessagingBannerDisplay *bannerMessage = [[FIRInAppMessagingBannerDisplay alloc]
         initWithMessageID:definition.renderData.messageID
              campaignName:definition.renderData.name
+        experimentPayload:definition.experimentPayload
       renderAsTestMessage:definition.isTestMessage
               triggerType:triggerType
                 titleText:title
@@ -491,6 +502,7 @@
   FIRInAppMessagingImageOnlyDisplay *imageOnlyMessage = [[FIRInAppMessagingImageOnlyDisplay alloc]
         initWithMessageID:definition.renderData.messageID
              campaignName:definition.renderData.name
+        experimentPayload:definition.experimentPayload
       renderAsTestMessage:definition.isTestMessage
               triggerType:triggerType
                 imageData:imageData
@@ -528,6 +540,7 @@
   FIRInAppMessagingModalDisplay *modalViewMessage = [[FIRInAppMessagingModalDisplay alloc]
         initWithMessageID:definition.renderData.messageID
              campaignName:definition.renderData.name
+        experimentPayload:definition.experimentPayload
       renderAsTestMessage:definition.isTestMessage
               triggerType:triggerType
                 titleText:title