+++ FirebaseABTesting/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v3.2.0
+- [added] Added completion handler for FIRExperimentController's updateExperimentsWithServiceOrigin method.
+- [deprecated] Deprecated `FIRExperimentController.updateExperiments(serviceOrigin:events:policy:lastStartTime:payloads:)`.
+- [added] Added method to expire experiments that are no longer running.
+- [added] Added method to directly activate an experiment.
+
 # v3.1.1
 - [fixed] Fixed an Analyzer issue (#3622).