+++ Firestore/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+# v1.11.1
+- [fixed] Firestore should now recover its connection to the server more
+  quickly after returning from the background (#4905).
+
 # v1.11.0
 - [changed] Improved performance of queries with large result sets.